+++ picopt/path.py
@@ -1,5 +1,6 @@
 """Data classes."""
 from dataclasses import dataclass
+from os import stat_result
 from pathlib import Path
 
 from confuse import AttrDict
@@ -14,6 +15,7 @@ class PathInfo:
     container_mtime: float | None
     convert: bool
     is_case_sensitive: bool
+    stat: stat_result | None = None
 
 
 def is_path_ignored(config: AttrDict, path: Path):